alt bro

+++ competra.wwwapi/Models/UserActivity.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public int UserId { get; set; }
-        public double Elo { get; set; }
+        public int Elo { get; set; }
         public int ActivityId { get; set; }
         public Activity Activity { get; set; }
         public User User { get; set; }
